refactor: Inline unused TransitFeedParameters
