+++ src/main/java/org/opentripplanner/standalone/config/feed/GtfsFeedParameters.java
@@ -1,12 +1,33 @@
 package org.opentripplanner.standalone.config.feed;
 
+import java.net.URI;
+import java.util.Objects;
+import java.util.Optional;
+import org.opentripplanner.graph_builder.model.DataSourceConfig;
+
 /**
  * Configure a GTFS feed.
  * Example: {@code [ {type="gtfs", source: "file:///path/to/otp/norway-gtfs.zip"} ] }
  */
-public class GtfsFeedParameters extends TransitFeedParameters {
+public class GtfsFeedParameters implements DataSourceConfig {
+
+  private final URI source;
+  private final String feedId;
 
   GtfsFeedParameters(GtfsFeedParametersBuilder builder) {
-    super(builder.source(), builder.feedId());
+    this.source = Objects.requireNonNull(builder.source());
+    this.feedId = builder.feedId();
+  }
+
+  @Override
+  public URI source() {
+    return source;
+  }
+
+  /**
+   * The unique ID for this feed.
+   */
+  public Optional<String> feedId() {
+    return Optional.ofNullable(feedId);
   }
 }
