+++ src/main/java/org/opentripplanner/standalone/config/feed/NetexFeedParameters.java
@@ -1,17 +1,19 @@
 package org.opentripplanner.standalone.config.feed;
 
+import java.net.URI;
+import java.util.Objects;
 import java.util.Optional;
 import java.util.regex.Pattern;
+import org.opentripplanner.graph_builder.model.DataSourceConfig;
 
 /**
  * Configure a NeTEx feed. Overrides default values specified in {@link NetexDefaultParameters}
  * Example: {@code [ {type="netex", source: "file:///path/to/otp/norway-netex.zip"} ] }
  */
-public class NetexFeedParameters extends TransitFeedParameters {
+public class NetexFeedParameters implements DataSourceConfig {
 
-  /**
-   *
-   */
+  private final URI source;
+  private final String feedId;
   private final Pattern sharedFilePattern;
 
   /**
@@ -30,11 +32,24 @@ public class NetexFeedParameters extends TransitFeedParameters {
   private final Pattern groupFilePattern;
 
   NetexFeedParameters(NetexFeedParametersBuilder builder) {
-    super(builder.source(), builder.feedId());
+    this.source = Objects.requireNonNull(builder.source());
+    this.feedId = builder.feedId();
     this.sharedFilePattern = builder.sharedFilePattern();
     this.sharedGroupFilePattern = builder.sharedGroupFilePattern();
     this.ignoreFilePattern = builder.ignoreFilePattern();
     this.groupFilePattern = builder.groupFilePattern();
+  }
+
+  @Override
+  public URI source() {
+    return source;
+  }
+
+  /**
+   * The unique ID for this feed.
+   */
+  public Optional<String> feedId() {
+    return Optional.ofNullable(feedId);
   }
 
   /**
